Some implementation, but the AU no longer works
In the example AU host project, it outputs silence. In AU Lab, it fails to load with error -10863 (not allowed in current context). In auval, it fails to validate with error -10874 (too many frames to process).

I’m also supposed to be setting some properties somewhere, but the documentation only explains how to query those properties as an AU host, not how to set them in an AU. Apple’s demo filter AU doesn’t compile because I don’t have an Apple developer account, so I can’t run auval on it to see whether it sets those properties correctly and then crib from its implementation.