Add noise gate
Now when RNNoise reports less than 95% confidence that the buffer is speech, the plugin replaces it with silence.

This adds two parameters to the AU: speechConfidenceThresholdPct and gateAttackDelay. They control how confident RNNoise must be about the speechiness of the buffer before the plugin doesn’t zero it, and the number of milliseconds of buffer to leave untouched, so that trail-offs don’t get cut out.